refactor: move abi and selectors code to common
