+++ cli/src/cmd/forge/fourbyte.rs
@@ -4,8 +4,10 @@ use crate::{
 };
 use clap::Parser;
 use ethers::prelude::artifacts::output_selection::ContractOutputSelection;
-use foundry_common::compile;
-use foundry_utils::selectors::{import_selectors, SelectorImportData};
+use foundry_common::{
+    compile,
+    selectors::{import_selectors, SelectorImportData},
+};
 
 #[derive(Debug, Clone, Parser)]
 pub struct UploadSelectorsArgs {
