+++ common/Cargo.toml
@@ -17,6 +17,7 @@ foundry-config = { path = "../config" }
 ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
 ethers-solc = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
 ethers-providers = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
+ethers-etherscan = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
 
 # io
 reqwest = { version = "0.11", default-features = false }
@@ -45,3 +46,6 @@ semver = "1.0.5"
 once_cell = "1.13.1"
 dunce = "1.0.2"
 regex = "1.6.0"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
