+++ common/src/lib.rs
@@ -1,6 +1,7 @@
 //! Common utilities for building and using foundry's tools.
 #![deny(missing_docs, unused_crate_dependencies)]
 
+pub mod abi;
 pub mod calc;
 pub mod clap_helpers;
 pub mod compile;
@@ -11,6 +12,7 @@ pub mod evm;
 pub mod fmt;
 pub mod fs;
 pub mod provider;
+pub mod selectors;
 pub mod shell;
 pub use provider::*;
 pub mod term;
