+++ evm/src/fuzz/mod.rs
@@ -276,7 +276,7 @@ impl BaseCounterExample {
 
 impl fmt::Display for BaseCounterExample {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let args = foundry_utils::format_tokens(&self.args).collect::<Vec<_>>().join(", ");
+        let args = foundry_common::abi::format_tokens(&self.args).collect::<Vec<_>>().join(", ");
 
         if let Some(sender) = self.sender {
             write!(f, "sender={:?} addr=", sender)?
